# gen/measurement/v1/measurement.openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "measurement/v1/measurement.proto",
    "version": "version not set"
  },
...
  "paths": {
    "/v1/customer/{customerId}/measurement": {
      "post": {
        "operationId": "MeasurementService_CreateMeasurement",
        "responses": {
          "200": {
            "description": "A successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/v1CreateMeasurementResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/rpcStatus"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeasurementServiceCreateMeasurementBody"
              }
            }
          },
          "required": true
        },
        "tags": ["MeasurementService"]
      }
    },
...
  },
  "components": {
    "schemas": {
      "MeasurementServiceCreateMeasurementBody": {
        "type": "object",
        "properties": {
          "measuredOn": {
            "$ref": "#/components/schemas/typeDate"
          },
          "measuredBy": {
            "type": "string"
          },
          "isDraft": {
            "type": "boolean"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/v1MeasurementEntry"
            }
          }
        }
      },
...
      "protobufAny": {
        "type": "object",
        "properties": {
          "@type": {
            "type": "string"
          }
        },
        "additionalProperties": {}
      },
      "rpcStatus": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/protobufAny"
            }
          }
        }
      },
      "typeDate": {
        "type": "object",
        "properties": {
          "year": {
            "type": "integer",
            "format": "int32",
            "description": "Year of the date. Must be from 1 to 9999, or 0 to specify a date without\na year."
          },
          "month": {
            "type": "integer",
            "format": "int32",
            "description": "Month of a year. Must be from 1 to 12, or 0 to specify a year without a\nmonth and day."
          },
          "day": {
            "type": "integer",
            "format": "int32",
            "description": "Day of a month. Must be from 1 to 31 and valid for the year and month, or 0\nto specify a year by itself or a year and month where the day isn't\nsignificant."
          }
        },
        "description": "* A full date, with non-zero year, month, and day values.\n* A month and day, with a zero year (for example, an anniversary).\n* A year on its own, with a zero month and a zero day.\n* A year and month, with a zero day (for example, a credit card expiration\n  date).\n\nRelated types:\n\n* [google.type.TimeOfDay][google.type.TimeOfDay]\n* [google.type.DateTime][google.type.DateTime]\n* [google.protobuf.Timestamp][google.protobuf.Timestamp]",
        "title": "Represents a whole or partial calendar date, such as a birthday. The time of\nday and time zone are either specified elsewhere or are insignificant. The\ndate is relative to the Gregorian Calendar. This can represent one of the\nfollowing:"
      },
      "v1CreateMeasurementResponse": {
        "type": "object",
        "properties": {
          "measurementId": {
            "type": "string"
          }
        }
      },
...
      "v1MeasurementEntry": {
        "type": "object",
        "properties": {
          "measurementItemId": {
            "type": "string"
          },
          "unmeasurable": {
            "type": "boolean"
          },
          "note": {
            "type": "string"
          },
          "values": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/v1MeasurementValue"
            }
          }
        }
      },
      "v1MeasurementValue": {
        "type": "object",
        "properties": {
          "trialIndex": {
            "type": "integer",
            "format": "int64"
          },
          "side": {
            "$ref": "#/components/schemas/v1Side"
          },
          "value": {
            "type": "number",
            "format": "double"
          },
          "valueSecondary": {
            "type": "number",
            "format": "double"
          },
          "valueChoice": {
            "type": "string"
          }
        }
      },
      "v1Side": {
        "type": "string",
        "enum": ["SIDE_UNSPECIFIED", "SIDE_NONE", "SIDE_LEFT", "SIDE_RIGHT"],
        "default": "SIDE_UNSPECIFIED"
      },
...
    }
  }
}